requirements.txt
langchain==0.1.0
langchain-community==0.0.10
chromadb==0.4.22
sentence-transformers==2.2.2
pypdf2==3.0.1
python-docx==1.1.0
pandas==2.1.4
numpy==1.24.3
streamlit==1.29.0
ollama==0.1.7
python-dotenv==1.0.0